backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
google-generativeai
python-dotenv
openai
